fix issue for delete operation and document debezium source connector issue
also mark feature experimental in readme since the _id field typing issue from debezium
can be a show stopper when trying to use CDC for documents with _id field other
than type string
